Modify searching/adding stocks to reference the portfolio from which that page was accessed.
This removes the most aggregious remaining hard-coded reference to the seeded portfolio.
(The root url still references the seeded portfolio for the landing page view,
which is intentional to unify the experience of users viewing the app for the first time.)

This change required moving a couple of global navigation buttons into non-global
templates/views becuse those buttons now require knowledge of the relevant portfolio.